delete workingsetstorecommon
